# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportConfirm": {
        "properties": {
          "confirmation": {
            "title": "Confirmation",
            "type": "boolean"
          }
        },
        "required": [
          "confirmation"
        ],
        "title": "ImportConfirm",
        "type": "object"
      },
...
      "ReceiptUpdate": {
        "properties": {
          "path": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "ReceiptUpdate",
        "type": "object"
      },
...
      "ToolDefinition": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "mutates": {
            "title": "Mutates",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "requires_confirmation": {
            "default": false,
            "title": "Requires Confirmation",
            "type": "boolean"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "name",
          "roles",
          "mutates",
          "description"
        ],
        "title": "ToolDefinition",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API de presupuesto, gastos, cierres y herramientas firmadas de Hermes.",
    "title": "D89 Sistema de Gastos API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/dashboard": {
      "get": {
        "operationId": "get_dashboard_api_v1_dashboard_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Dashboard Api V1 Dashboard Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dashboard",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/expenses": {
      "get": {
        "operationId": "get_expenses_api_v1_expenses_get",
        "parameters": [
          {
            "in": "query",
            "name": "work_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Expenses Api V1 Expenses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Expenses",
        "tags": [
          "expenses"
        ]
      },
      "post": {
        "operationId": "post_expense_api_v1_expenses_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Expense Api V1 Expenses Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Expense",
        "tags": [
          "expenses"
        ]
      }
    },
    "/api/v1/expenses/{expense_id}/receipt": {
      "patch": {
        "operationId": "patch_expense_receipt_api_v1_expenses__expense_id__receipt_patch",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Expense Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Patch Expense Receipt Api V1 Expenses  Expense Id  Receipt Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Expense Receipt",
        "tags": [
          "expenses"
        ]
      }
    },
...
    "/api/v1/hermes/hmac-probe": {
      "post": {
        "operationId": "hmac_probe_api_v1_hermes_hmac_probe_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-d89-key-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Key-Id"
            }
          },
          {
            "in": "header",
            "name": "x-d89-timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "x-d89-nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Nonce"
            }
          },
          {
            "in": "header",
            "name": "x-d89-signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Hmac Probe Api V1 Hermes Hmac Probe Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hmac Probe",
        "tags": [
          "hermes"
        ]
      }
    },
    "/api/v1/hermes/tools": {
      "get": {
        "operationId": "hermes_tools_api_v1_hermes_tools_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-d89-key-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Key-Id"
            }
          },
          {
            "in": "header",
            "name": "x-d89-timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "x-d89-nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Nonce"
            }
          },
          {
            "in": "header",
            "name": "x-d89-signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ToolDefinition"
                  },
                  "title": "Response Hermes Tools Api V1 Hermes Tools Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hermes Tools",
        "tags": [
          "hermes"
        ]
      }
    },
    "/api/v1/hermes/tools/{tool_name}": {
      "post": {
        "operationId": "invoke_hermes_tool_api_v1_hermes_tools__tool_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "tool_name",
            "required": true,
            "schema": {
              "title": "Tool Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-d89-actor-phone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Actor-Phone"
            }
          },
          {
            "in": "header",
            "name": "x-d89-key-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Key-Id"
            }
          },
          {
            "in": "header",
            "name": "x-d89-timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "x-d89-nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Nonce"
            }
          },
          {
            "in": "header",
            "name": "x-d89-signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-D89-Signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Invoke Hermes Tool Api V1 Hermes Tools  Tool Name  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invoke Hermes Tool",
        "tags": [
          "hermes"
        ]
      }
    },
    "/api/v1/incomes": {
      "get": {
        "operationId": "get_incomes_api_v1_incomes_get",
        "parameters": [
          {
            "in": "query",
            "name": "work_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Incomes Api V1 Incomes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incomes",
        "tags": [
          "cashflow"
        ]
      },
      "post": {
        "operationId": "post_income_api_v1_incomes_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Income Api V1 Incomes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Income",
        "tags": [
          "cashflow"
        ]
      }
    },
    "/api/v1/neodata/imports/{import_id}/confirm": {
      "post": {
        "operationId": "confirm_neodata_import_api_v1_neodata_imports__import_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Confirm Neodata Import Api V1 Neodata Imports  Import Id  Confirm Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Neodata Import",
        "tags": [
          "neodata"
        ]
      }
    },
    "/api/v1/neodata/preview": {
      "post": {
        "operationId": "preview_neodata_api_v1_neodata_preview_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_neodata_api_v1_neodata_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Neodata Api V1 Neodata Preview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Neodata",
        "tags": [
          "neodata"
        ]
      }
    },
    "/api/v1/reports/works/{work_id}.pdf": {
      "get": {
        "operationId": "export_work_pdf_api_v1_reports_works__work_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Work Pdf",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/works/{work_id}.xlsx": {
      "get": {
        "operationId": "export_work_excel_api_v1_reports_works__work_id__xlsx_get",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Work Excel",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}